cosmetic changes, flow

+++ scripts/ptest.pl
@@ -5,7 +5,6 @@
 #system("screen -S demo -X stuff 'date\n'");
 #system("screen -S demo -X stuff 'ls -ltr\n'");
 
-
 ###Get Hostname###
 $Host=`hostname`;
 chomp($Host);
@@ -18,11 +17,13 @@ if($a=~ /(.*).*SN(.*).*SKU(.*)/){
     $SKU=$3;
     $DIR=substr($MACID,9,8);
     $DIR=~ s/\://g;
-    system("rm -r Logs/$DIR");
+    if(-e "Logs/$DIR"){
+	system("rm -r Logs/$DIR");
+    }
 }
 
 ###POWER ON BOARD###
-print "2.) Connect cables and power on board, press enter when done: <enter>\n";
+print "2.) Boot the board WITHOUT the SD card inserted, press enter when done: <enter>";
 $a = <STDIN>;
 
 ###FLASH THE BOARD###
@@ -30,11 +31,13 @@ system("screen -s para -X stuff 'printenv\n'");
 system("screen -s para -X stuff 'setenv ethaddr $MACID\n'");
 system("screen -s para -X stuff 'setenv AdaptevaSKU $SKU\n'");
 system("screen -s para -X stuff 'saveenv\n'");
-print "3.) The board is being programmed, wait for \"done\" and the \"zynq-uboot\" prompt in monitor window:<enter>\n";
+print "3.) Flashing the board, wait for \"done\" & prompt in 2nd window:<enter>";
 $a = <STDIN>;
 
 ###REBOOT###
-print "4.) Turn off power-->insert SD card,-->reapply power, press enter when you see \"root@linaro-nano:~#\" prompt in monitor window:<enter>";
+print "4.) Reboot the board WITH the SD card inserted and press enter:<enter>";
+$a = <STDIN>;
+print "5.) Press enter when you see \"root\@linaro-nano:~#\" prompt in 2nd window:<enter>";
 $a = <STDIN>;
 
 ###RUN TEST###
@@ -42,7 +45,7 @@ system("screen -S para -X stuff 'su linaro\n cd\n run\n'");
 system("screen -S para -X stuff 'rcp -r $DIR $Host:Logs\n'");
 
 ###CHECK FOR PASS###
-print "7.) Wait until test completes in monitor window:<enter>";
+print "5.) Waiting for test to complete...\n";
 
 ###DISPLAY PASS/FAIL###
 while(1){
